Remove Feedback tab from ShellPage and add to Settings
