+++ settings/DevHome.Settings/Extensions/ServiceExtensions.cs
@@ -12,6 +12,8 @@ public static class ServiceExtensions
     public static IServiceCollection AddSettings(this IServiceCollection services, HostBuilderContext context)
     {
         // Project services
+        services.AddTransient<FeedbackViewModel>();
+        services.AddTransient<FeedbackPage>();
         services.AddTransient<SettingsViewModel>();
         services.AddTransient<SettingsPage>();
         services.AddTransient<AboutViewModel>();
