+++ settings/DevHome.Settings/ViewModels/FeedbackViewModel.cs
@@ -1,0 +1,65 @@
+﻿// Copyright (c) Microsoft Corporation and Contributors
+// Licensed under the MIT license.
+
+using System.Reflection;
+using System.Windows.Input;
+using CommunityToolkit.Mvvm.ComponentModel;
+using CommunityToolkit.Mvvm.Input;
+using DevHome.Common.Extensions;
+using DevHome.Common.Helpers;
+using DevHome.Common.Services;
+using DevHome.Contracts.Services;
+using Microsoft.UI.Xaml;
+using Windows.ApplicationModel;
+
+namespace DevHome.Settings.ViewModels;
+
+public class FeedbackViewModel : ObservableRecipient
+{
+    private readonly IThemeSelectorService _themeSelectorService;
+    private ElementTheme _elementTheme;
+    private string _versionDescription;
+
+    public ElementTheme ElementTheme
+    {
+        get => _elementTheme;
+        set => SetProperty(ref _elementTheme, value);
+    }
+
+    public string VersionDescription
+    {
+        get => _versionDescription;
+        set => SetProperty(ref _versionDescription, value);
+    }
+
+    public ICommand SwitchThemeCommand
+    {
+        get;
+    }
+
+    public FeedbackViewModel(IThemeSelectorService themeSelectorService)
+    {
+        _themeSelectorService = themeSelectorService;
+        _elementTheme = _themeSelectorService.Theme;
+        _versionDescription = GetVersionDescription();
+
+        SwitchThemeCommand = new RelayCommand<ElementTheme>(
+            async (param) =>
+            {
+                if (ElementTheme != param)
+                {
+                    ElementTheme = param;
+                    await _themeSelectorService.SetThemeAsync(param);
+                }
+            });
+    }
+
+    private static string GetVersionDescription()
+    {
+        IAppInfoService appInfoService = Application.Current.GetService<IAppInfoService>();
+        var localizedAppName = appInfoService.GetAppNameLocalized();
+        var version = appInfoService.GetAppVersion();
+
+        return $"{localizedAppName} - {version.Major}.{version.Minor}.{version.Build}.{version.Revision}";
+    }
+}
