+++ settings/DevHome.Settings/ViewModels/SettingsViewModel.cs
@@ -55,6 +55,7 @@ public partial class SettingsViewModel : ObservableRecipient
             new Setting("Accounts", string.Empty, stringResource.GetLocalized("Settings_Accounts_Header"), stringResource.GetLocalized("Settings_Accounts_Description"), "\ue77b", false),
             new Setting("Extensions", string.Empty, stringResource.GetLocalized("Settings_Extensions_Header"), stringResource.GetLocalized("Settings_Extensions_Description"), "\ued35", false),
             new Setting("About", string.Empty, stringResource.GetLocalized("Settings_About_Header"), stringResource.GetLocalized("Settings_About_Description"), "\ue946", false),
+            new Setting("Feedback", string.Empty, stringResource.GetLocalized("Settings_Feedback_Header"), stringResource.GetLocalized("Settings_Feedback_Description"), "\ue939", false),
         };
 
         foreach (var setting in settings)
@@ -80,6 +81,9 @@ public partial class SettingsViewModel : ObservableRecipient
                 return;
             case "About":
                 navigationService.NavigateTo(typeof(AboutViewModel).FullName!);
+                return;
+            case "Feedback":
+                navigationService.NavigateTo(typeof(FeedbackViewModel).FullName!);
                 return;
             default:
                 return;
